Use ESP UUID provided directly by sd-boot
Rather than discovering the ESP's UUID using efibootmgr, the bootbind
script can directly obtain the UUID stored in the EFI variables by
systemd-boot. This eliminates any possibility of getting the wrong
UUID, and also works properly if systemd-boot was loaded as a second
(or third...) bootloader by the system's built-in EFI firmware.